Added helper scripts + made text changes.
This was quite a big commit... as well as adding a helper mongo script to fill in some data (more to be added).
The text on the page was tided so it doesn't look quite as terrible.  Also the search results look a little cleaner, although
there is still a ton of work left to do on it.